rename dir to tag-input
